+++ src/resfinder/cge/pointfinder.py
@@ -411,7 +411,7 @@ class PointFinder(CGEFinder):
         known = []
         for mis_match in mis_matches:
             mis_match_key = (f"{gene_db_id}_{mis_match[1]}"
-                             f"_{mis_match[8].lower()}")
+                             f"_{mis_match[-1].lower()}")
             if mis_match_key in phenodb:
                 known.append(mis_match)
         return known