+++ CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 4.2.1 - 2022-09-12
+
+## [4.2.2] - 2022-09-19
+
+### Added
+- ResFinder will now complain and exit if the ResFinder database is not found, as it is necessary, even if only looking for point mutations or disinfectant genes.
+
+### Fixed
+- Issue where the application failed when run using only the pointfinder option (--point)
+- Changelog version format from d.d.d to [d.d.d]
+
+
+## [4.2.1] - 2022-09-12
 
 ### Added
 - Several environmental variables for ResFinder settings (see README.md).
@@ -39,4 +50,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Issue in json ouptput where unknown mutations were listed even if option wasn't enabled.
 
 
-## 4.2.0 - 2022-04-21 [YANKED]
+## [4.2.0] - 2022-04-21 [YANKED]